# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export interface TelegramAuthInput {
  /** @minLength 1 */
  initData: string;
}

export interface TelegramWebhookUpdate { [key: string]: unknown }

export interface UserProfile {
  id: string;
  telegramUserId: string;
  /** @nullable */
  username: string | null;
  firstName: string;
  /** @nullable */
  lastName?: string | null;
  /** @nullable */
  displayName?: string | null;
  /** @nullable */
  photoUrl?: string | null;
  /** @nullable */
  languageCode?: string | null;
  isPremium?: boolean;
  balance: number;
  createdAt: string;
}

export interface MinerConfig {
  level: number;
  name: string;
  miningRate: number;
  maxEnergyMinutes: number;
  upgradeCost: number;
}

export interface MinerState {
  level: number;
  name: string;
  miningRate: number;
  maxEnergyMinutes: number;
  currentEnergyMinutes: number;
  minedAmount: number;
  balance: number;
  isMining: boolean;
  lastUpdatedAt: string;
  nextUpgrade: MinerConfig | null;
}

export interface AppSession {
  user: UserProfile;
  miner: MinerState;
}

export interface ClaimResult {
  claimedAmount: number;
  balance: number;
  miner: MinerState;
}

export interface MinerUpgradeInput {
  /**
     * @minimum 2
     * @maximum 3
     */
  targetLevel: number;
}

export interface UpgradeResult {
  balance: number;
  miner: MinerState;
}

export type AdRewardIntentProvider = typeof AdRewardIntentProvider[keyof typeof AdRewardIntentProvider];


export const AdRewardIntentProvider = {
  monetag: 'monetag',
} as const;

export interface AdRewardIntent {
  intentId: string;
  rewardMinutes: number;
  provider: AdRewardIntentProvider;
  expiresAt: string;
}

export interface AdRewardInput {
  /** @minLength 1 */
  intentId: string;
}

export interface RewardResult {
  rewardMinutes: number;
  miner: MinerState;
}

export type ActivityItemType = typeof ActivityItemType[keyof typeof ActivityItemType];


export const ActivityItemType = {
  MINING_CLAIM: 'MINING_CLAIM',
  AD_REWARD: 'AD_REWARD',
  AD_REWARD_LP: 'AD_REWARD_LP',
  MINER_UPGRADE: 'MINER_UPGRADE',
  PURCHASE: 'PURCHASE',
  REVEAL_UNLOCK: 'REVEAL_UNLOCK',
} as const;

export interface ActivityItem {
  id: string;
  type: ActivityItemType;
  amount: number;
  label: string;
  createdAt: string;
}

export type SocialDataRevealProgress = {[key: string]: number};

export type SocialDataViewCounts = {[key: string]: number};

/**
 * Canonical social/gameplay aggregate owned by the backend.
 */
export interface SocialData {
  starterIds?: string[];
  unlockedIds?: string[];
  revealProgress?: SocialDataRevealProgress;
  lpBalance?: number;
  followingIds?: string[];
  likedPostIds?: string[];
  hotPostIds?: string[];
  savedPostIds?: string[];
  followedStyles?: string[];
  muted?: boolean;
  viewCounts?: SocialDataViewCounts;
  ownedMiners?: string[];
  ownedSkins?: string[];
  equippedMiner?: string | null;
  equippedSkin?: string | null;
  modelTickets?: number;
  shotTickets?: number;
  bonusMineMinutes?: number;
  lastWheelSpin?: string | null;
  earnedSpins?: number;
  adsWatchedForSpin?: number;
}

export type WheelPrizeKind = typeof WheelPrizeKind[keyof typeof WheelPrizeKind];


export const WheelPrizeKind = {
  rm: 'rm',
  lp: 'lp',
  modelTicket: 'modelTicket',
  shotTicket: 'shotTicket',
  mineMinutes: 'mineMinutes',
} as const;

export interface WheelPrize {
  key: string;
  label: string;
  kind: WheelPrizeKind;
  value: number;
}

export type GetSocial200 = {
  data?: SocialData;
};

export type RevealTapBody = {
  modelId: string;
};

export type RevealTap200 = {
  data?: SocialData;
};

export type RevealUnlockBody = {
  modelId: string;
};

export type RevealUnlock200 = {
  data?: SocialData;
};

export type FollowModelBody = {
  modelId: string;
};

export type FollowModel200 = {
  data?: SocialData;
};

export type PostInteractionBodyFlag = typeof PostInteractionBodyFlag[keyof typeof PostInteractionBodyFlag];


export const PostInteractionBodyFlag = {
  liked: 'liked',
  saved: 'saved',
  hot: 'hot',
} as const;

export type PostInteractionBody = {
  postId: string;
  flag: PostInteractionBodyFlag;
  value: boolean;
};

export type PostInteraction200 = {
  data?: SocialData;
};

export type RecordViewBody = {
  postId: string;
};

export type RecordView200 = {
  data?: SocialData;
};

export type FollowStyleBody = {
  styleId: string;
};

export type SetMuteBody = {
  value: boolean;
};

export type WheelSpin200Prize = { [key: string]: unknown };

export type WheelSpin200 = {
  prize?: WheelSpin200Prize;
  data?: SocialData;
};

export type PurchaseItemBody = {
  itemId: string;
};

export type EquipItemBody = {
  itemId: string;
};