# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/telegram:
    post:
      operationId: authenticateTelegram
      tags: [auth]
      summary: Authenticate with Telegram Mini App initData
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TelegramAuthInput"
      responses:
        "200":
          description: Authenticated user and miner state
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AppSession"
        "401":
          description: Invalid or expired Telegram initData
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /auth/preview:
    post:
      operationId: authenticatePreview
      tags: [auth]
      summary: Authenticate a browser preview session
      description: Creates a development-only preview account for direct browser access. Disabled in production.
      responses:
        "200":
          description: Authenticated preview user and miner state
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AppSession"
        "403":
          description: Browser preview authentication is disabled
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /me:
    get:
      operationId: getMe
      tags: [auth]
      summary: Get the authenticated RM Coin account
      security:
        - sessionCookie: []
      responses:
        "200":
          description: Account details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserProfile"
        "401":
          description: Authentication required
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /miner:
    get:
      operationId: getMiner
      tags: [mining]
      summary: Get server-calculated miner state
      security:
        - sessionCookie: []
      responses:
        "200":
          description: Current miner state
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MinerState"
        "401":
          description: Authentication required
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /miner/claim:
    post:
      operationId: claimMining
      tags: [mining]
      summary: Atomically extract unclaimed RM Coin
      security:
        - sessionCookie: []
      responses:
        "200":
          description: Updated balance and miner state
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ClaimResult"
        "401":
          description: Authentication required
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /miner/upgrade:
    post:
      operationId: upgradeMiner
      tags: [mining]
      summary: Upgrade to the next miner level
      security:
        - sessionCookie: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MinerUpgradeInput"
      responses:
        "200":
          description: Updated miner state and balance
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UpgradeResult"
        "400":
          description: Upgrade is not available or balance is insufficient
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /ads/reward-intent:
    post:
      operationId: createAdRewardIntent
      tags: [ads]
      summary: Create a one-time Monetag reward intent
      security:
        - sessionCookie: []
      responses:
        "200":
          description: Intent to be completed after a successful Monetag ad
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdRewardIntent"
  /ads/reward:
    post:
      operationId: completeAdReward
      tags: [ads]
      summary: Complete a Monetag rewarded ad and grant energy
      security:
        - sessionCookie: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AdRewardInput"
      responses:
        "200":
          description: Updated miner state after reward
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RewardResult"
        "409":
          description: Intent already used or expired
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /activity:
    get:
      operationId: getActivity
      tags: [activity]
      summary: Get recent mining and reward activity
      security:
        - sessionCookie: []
      responses:
        "200":
          description: Recent activity
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ActivityItem"
...
components:
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: rm_session
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required: [error]
    TelegramAuthInput:
      type: object
      properties:
        initData:
          type: string
          minLength: 1
      required: [initData]
...
    UserProfile:
      type: object
      properties:
        id: { type: string }
        telegramUserId: { type: string }
        username: { type: ["string", "null"] }
        firstName: { type: string }
        lastName: { type: ["string", "null"] }
        displayName: { type: ["string", "null"] }
        photoUrl: { type: ["string", "null"] }
        languageCode: { type: ["string", "null"] }
        isPremium: { type: boolean }
        balance: { type: number }
        createdAt: { type: string, format: date-time }
      required: [id, telegramUserId, username, firstName, balance, createdAt]
    MinerConfig:
      type: object
      properties:
        level: { type: number }
        name: { type: string }
        miningRate: { type: number }
        maxEnergyMinutes: { type: number }
        upgradeCost: { type: number }
      required: [level, name, miningRate, maxEnergyMinutes, upgradeCost]
    MinerState:
      type: object
      properties:
        level: { type: number }
        name: { type: string }
        miningRate: { type: number }
        maxEnergyMinutes: { type: number }
        currentEnergyMinutes: { type: number }
        minedAmount: { type: number }
        balance: { type: number }
        isMining: { type: boolean }
        lastUpdatedAt: { type: string, format: date-time }
        nextUpgrade:
          anyOf:
            - $ref: "#/components/schemas/MinerConfig"
            - type: "null"
      required: [level, name, miningRate, maxEnergyMinutes, currentEnergyMinutes, minedAmount, balance, isMining, lastUpdatedAt, nextUpgrade]
    AppSession:
      type: object
      properties:
        user:
          $ref: "#/components/schemas/UserProfile"
        miner:
          $ref: "#/components/schemas/MinerState"
      required: [user, miner]
    ClaimResult:
      type: object
      properties:
        claimedAmount: { type: number }
        balance: { type: number }
        miner:
          $ref: "#/components/schemas/MinerState"
      required: [claimedAmount, balance, miner]
    MinerUpgradeInput:
      type: object
      properties:
        targetLevel: { type: number, minimum: 2, maximum: 3 }
      required: [targetLevel]
    UpgradeResult:
      type: object
      properties:
        balance: { type: number }
        miner:
          $ref: "#/components/schemas/MinerState"
      required: [balance, miner]
    AdRewardIntent:
      type: object
      properties:
        intentId: { type: string }
        rewardMinutes: { type: number }
        provider: { type: string, enum: [monetag] }
        expiresAt: { type: string, format: date-time }
      required: [intentId, rewardMinutes, provider, expiresAt]
    AdRewardInput:
      type: object
      properties:
        intentId: { type: string, minLength: 1 }
      required: [intentId]
    RewardResult:
      type: object
      properties:
        rewardMinutes: { type: number }
        miner:
          $ref: "#/components/schemas/MinerState"
      required: [rewardMinutes, miner]
    ActivityItem:
      type: object
      properties:
        id: { type: string }
        type: { type: string, enum: [MINING_CLAIM, AD_REWARD, AD_REWARD_LP, MINER_UPGRADE, PURCHASE, REVEAL_UNLOCK] }
        amount: { type: number }
        label: { type: string }
        createdAt: { type: string, format: date-time }
      required: [id, type, amount, label, createdAt]
...